+++ changelog.md
@@ -1,6 +1,14 @@
 # SFML Sound System Change Log
 All notable changes to this project will be documented in this file.
 
+## 1.2.0 - 2018-03-27
+### Added
+- Ability to manually set/get maximum number of voices allowed
+- Ability to specify sound location using only 2D co-ordinates
+- Ability to load a buffer from a memory block using char*
+### Fixed
+- noexcept (Visual Studio version mismatch and now unique ID)
+
 ## [1.1.0] - 2017-07-02
 ### Added
 - Ability to get a direct reference to a stored sound buffer